pyproject.toml
[tool.poetry]
name = "ls-1x"
version = "0.1.0"
description = "ls-1x site"
authors = ["Trieu Truong"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
schedule = "^1.2.1"
streamlit = "^1.34.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
